+++ raiden-dapp/tests/unit/views/backup-state.spec.ts
@@ -2,6 +2,8 @@ jest.useFakeTimers();
 import { mount, Wrapper } from '@vue/test-utils';
 import Vue from 'vue';
 import Vuetify from 'vuetify';
+
+import store from '@/store/index';
 import BackupState from '@/views/BackupState.vue';
 
 Vue.use(Vuetify);
@@ -18,7 +20,8 @@ describe('BackupState.vue', () => {
       stubs: ['v-dialog'],
       mocks: {
         $t: (msg: string) => msg
-      }
+      },
+      store
     });
   });
 
@@ -54,6 +57,19 @@ describe('BackupState.vue', () => {
 
   test('clicking upload state button opens upload state dialog', () => {
     expect(wrapper.vm.$data.uploadState).toBe(false);
+
+  test('download state button disabled if disconnected', () => {
+    const downloadStateButton = wrapper.find(
+      '.backup-state__buttons__download-state'
+    );
+
+    expect(downloadStateButton.classes()).toContain('v-list-item--disabled');
+  });
+
+  // test('upload state', () => {
+  //   const uploadStateTitle = wrapper.find(
+  //     '.backup-state__buttons__upload-state__title'
+  //   );
 
     const uploadStateButton = wrapper.find(
       '.backup-state__buttons__upload-state'